Add project, job and dashboard pages and style
Create a basic project and job model, controller and view. Very simple,
just a title and description for project, and a command for job. The two
are connected by a simple join table.

Dashboard page as the default index page, using the side menu layout
from PureCSS.

Style is from http://purecss.io/